separate code into 3 files
